requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements-dev.txt requirements-dev.in
#
bleach==4.1.0
    # via readme-renderer
certifi==2022.6.15
    # via requests
cffi==1.15.0
    # via cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.12
    # via requests
commonmark==0.9.1
    # via rich
cryptography==37.0.2
    # via secretstorage
distlib==0.3.4
    # via virtualenv
docutils==0.18.1
    # via readme-renderer
filelock==3.6.0
    # via virtualenv
flake8==4.0.1
    # via -r requirements-dev.in
identify==2.4.12
    # via pre-commit
idna==3.3
    # via requests
importlib-metadata==4.11.3
    # via
    #   keyring
    #   twine
invoke==1.7.1
    # via -r requirements-dev.in
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
keyring==23.5.0
    # via twine
marshmallow==2.21.0
    # via
    #   -r requirements-dev.in
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==0.23.1
    # via -r requirements-dev.in
mccabe==0.6.1
    # via flake8
nodeenv==1.6.0
    # via
    #   -r requirements-dev.in
    #   pre-commit
packaging==21.3
    # via bleach
pkginfo==1.8.2
    # via twine
platformdirs==2.5.1
    # via virtualenv
pre-commit==2.19.0
    # via -r requirements-dev.in
pycodestyle==2.8.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==2.4.0
    # via flake8
pygments==2.11.2
    # via
    #   readme-renderer
    #   rich
pyparsing==3.0.7
    # via packaging
pyyaml==6.0
    # via
    #   -r requirements-dev.in
    #   pre-commit
readme-renderer==35.0
    # via twine
requests==2.28.0
    # via
    #   requests-toolbelt
    #   twine
requests-toolbelt==0.9.1
    # via twine
rfc3986==2.0.0
    # via twine
rich==12.1.0
    # via twine
secretstorage==3.3.1
    # via keyring
six==1.16.0
    # via
    #   bleach
    #   virtualenv
sqlalchemy==1.3.24
    # via
    #   -r requirements-dev.in
    #   marshmallow-sqlalchemy
toml==0.10.2
    # via pre-commit
twine==4.0.1
    # via -r requirements-dev.in
urllib3==1.26.9
    # via
    #   requests
    #   twine
virtualenv==20.14.0
    # via pre-commit
webencodings==0.5.1
    # via bleach
zipp==3.7.0
    # via importlib-metadata
google-cloud-compute==1.5.1
    # via -r requirements.in
